feat/refactor: configure the LoginScreen UI to accessible UI and refactor the authentication form
